some change and fix

+++ blog/weblog/models.py
@@ -7,7 +7,7 @@ class User(AbstractUser):
         permissions = [('can_write','can write post'),
                        ('can_edit','can accept and edit other posts and comments'),
                        ('can_manage','can manage all content and users')]
-    phone = models.PositiveIntegerField("شماره",null=True)
+    phone = models.BigIntegerField("شماره",null=True)
     image = models.ImageField("تصویر", upload_to='users',null=True)
